+++ static/index.html
@@ -1,0 +1,75 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
+
+		<link rel="icon" href="assets/favicon.ico" />
+		<title>npm | build amazing things</title>
+
+		<link rel="stylesheet" href="styles/global.css" />
+	</head>
+	<body class="bg-light">
+		<!-- <input type="button" class="btn lead skew-fill red" value="See Plans"></input>
+        <input type="button" class="btn lead white" value="Join for free"></input>
+        <input type="button" class="btn lead skew-fill black wide" value="Create an Org"> -->
+
+		<header class="bg-white shadow-sm">
+			<div class="header-wraper gradient-border"></div>
+			<div class="header-wraper border-b-sm">
+				<div class="header-container my-5">
+					<div class="npm-expansions">
+						<span class="icon-sm mr-7 link">❤</span>
+						<p id="npm-expansions-txt" class="pointer link"></p>
+					</div>
+					<nav>
+						<ul>
+							<li class="mr-5 link">
+								<a href="">npm Enterprise</a>
+							</li>
+							<li class="mr-5 link"><a href="">Products</a></li>
+							<li class="mr-5 link"><a href="">Solutions</a></li>
+							<li class="mr-5 link"><a href="">Resources</a></li>
+							<li class="mr-5 link"><a href="">Docs</a></li>
+							<li class="mr-5 link"><a href="">Support</a></li>
+						</ul>
+					</nav>
+				</div>
+			</div>
+			<div class="header-wraper border-b-sm">
+				<div class="header-container">
+					<div class="brand mr-3">
+						<a href="">
+							<svg viewBox="0 0 780 250">
+								<path
+									fill="#231F20"
+									d="M240,250h100v-50h100V0H240V250z M340,50h50v100h-50V50z M480,0v200h100V50h50v150h50V50h50v150h50V0H480z M0,200h100V50h50v150h50V0H0V200z"
+								></path>
+							</svg>
+						</a>
+					</div>
+					<div class="input-group m-2">
+						<div class="input-group-prepend">
+							<img
+								class="icon-lg"
+								src="assets/magnifying-glass.png"
+							/>
+						</div>
+						<div class="input-group-area">
+							<input id="search-npm-packages" type="text" placeholder="Search Packages"/>
+						</div>
+					</div>
+					<div class="search-results-container">
+						<div class="top-result">
+							<p>Lorem Ipsum Dolor</p>
+							<p>Lorem Ipsum Dolor</p>
+						</div>
+					</div>
+				</div>
+			</div>
+		</header>
+
+		<script src="main.js"></script>
+	</body>
+</html>
